Add pagination utilities, slightly refactor response utilities
